feat: [USER STORY #81549] create rtcStatsPresenter in webrtc_connector

+++ apps/sender/lib/model/rtc_stats.dart
@@ -6,7 +6,7 @@ class RtcVideoOutboundStats {
   String? contentType;
   String? qualityLimitationReason;
   int? pliCount;
-  int? targetBitrate;
+  double? targetBitrate;
   double? encodeTime;
   bool? powerEfficientEncoder;
